Rename enhance to enhanceMachine
